Image metadata as a component.
